refactor(core): simplify selector parsing primitives
Split the token-parsing grab-bag by grammar responsibility — keys
(SelectorKeyParsing), numbers and booleans (SelectorNumberParsing),
tokens and quoted strings (SelectorStringParsing) — and rename
SelectorSnbtValidator/SelectorArgumentParser to the package's
uniform ...Parsing suffix.

Extract the one SNBT container loop (readSnbtElements) that compounds,
lists, and typed arrays each hand-rolled, and name the scalar
terminator predicate detekt flagged. readQuotedString's entry guard
becomes checkNotNull: every caller peeks the quote first, so a
non-quote start is a caller bug, not user parse input.

Behaviour and diagnostic offsets are pinned unchanged by the SNBT
characterization tests.
